requirements.txt
alembic==1.13.1
beautifulsoup4==4.12.3
feedfinder2==0.0.4
feedparser==6.0.11
flask_login==0.6.3
flask_migrate==4.0.7
pytest==8.2.2
python_dateutil==2.9.0.post0
pytz==2024.1
Requests==2.32.3
waitress==3.0.0
humanize==4.9.0
groq==0.8.0
